rename: 서비스 레이어 코로케이션 적용
